--- src/main/java/com/org/enums/Context.java
@@ -1,5 +1,0 @@
-package com.org.enums;
-
-public enum Context {
-	EXPECTED_DATA,DATA_TABLE, PAGE_DATA,SEARCH_DATA,FLAGS,TEMP_DATA_TABLE
-}
