--- src/main/java/com/org/enums/PlatformType.java
@@ -1,5 +1,0 @@
-package com.org.enums;
-
-public enum PlatformType {
-	LOCAL, REMOTE
-}
