--- src/main/java/com/org/enums/StoreValues.java
@@ -1,5 +1,0 @@
-package com.org.enums;
-
-public enum StoreValues {
-	
-}
