+++ src/main/java/com/org/generic/Enums/Context.java
@@ -1,0 +1,5 @@
+package com.org.generic.Enums;
+
+public enum Context {
+	EXPECTED_DATA,DATA_TABLE, PAGE_DATA,SEARCH_DATA,FLAGS,TEMP_DATA_TABLE
+}
