+++ src/main/java/com/org/generic/Enums/PlatformType.java
@@ -1,0 +1,5 @@
+package com.org.generic.Enums;
+
+public enum PlatformType {
+	LOCAL, REMOTE
+}
