+++ src/main/java/com/org/generic/Enums/StoreValues.java
@@ -1,0 +1,5 @@
+package com.org.generic.Enums;
+
+public enum StoreValues {
+	
+}
