+++ src/main/java/com/org/page/Common/CommonPage.java
@@ -1,0 +1,12 @@
+package com.org.page.Common;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.PageFactory;
+import org.openqa.selenium.support.pagefactory.AjaxElementLocatorFactory;
+
+public class CommonPage {
+	public CommonPage(WebDriver webDriver) {
+        PageFactory.initElements(new AjaxElementLocatorFactory(webDriver, 15), this);
+    }
+
+}
