--- src/main/java/com/org/pageClass/CommonClass/CommonPage.java
@@ -1,12 +1,0 @@
-package com.org.pageClass.CommonClass;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.PageFactory;
-import org.openqa.selenium.support.pagefactory.AjaxElementLocatorFactory;
-
-public class CommonPage {
-	public CommonPage(WebDriver webDriver) {
-        PageFactory.initElements(new AjaxElementLocatorFactory(webDriver, 15), this);
-    }
-
-}
